requirements.txt
pyspark
kafka-python
scikit-learn
joblib
pandas
numpy
